Task6-7

+++ Assets/Content/Features/AIModule/Scripts/Entity/EntityData.cs
@@ -4,8 +4,8 @@ namespace Content.Features.AIModule.Scripts.Entity {
     [Serializable]
     public class EntityData {
         public EntityType EntityType;
-        public int StartHealth;
-        public int Damage;
+        public float StartHealth;
+        public float Damage;
         public float AttackDistance;
         public float InteractDistance;
         public float Speed;
